+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - 新增主键实体IntegerIdKey。
 - 添加KeyFetcher的UUID以及DenseUUID实现。
+- 添加 PagingUtil.subList方法。
+- 完善RelationDao以及RelationService及其实现。
 
 #### Bug修复
 
@@ -13,7 +15,7 @@
 
 #### 功能移除
 
-- (无)
+- ~~移除PagingUtil中的indexBound方法。~~
 
 ---